+++ django_cryptolock/forms.py
@@ -1,9 +1,14 @@
 from django import forms
 from django.contrib.auth import authenticate
+from django.contrib.auth import get_user_model
+from django.core.exceptions import ValidationError
 from django.utils.translation import gettext, gettext_lazy as _
+from django.conf import settings
+
+from pybitid import bitid
 
 from .models import Address
-from .validators import validate_monero_address
+from .validators import validate_monero_address, validate_bitcoin_address
 from .utils import generate_challenge
 
 
@@ -17,7 +22,9 @@ class ChallengeMixin(forms.Form):
     challenge = forms.CharField()
 
     def include_challange(self):
-        new_challenge = generate_challenge()
+        new_challenge = bitid.build_uri(
+            self.request.build_absolute_uri(), generate_challenge()
+        )
         if not self.data:
             self.request.session["current_challenge"] = new_challenge
             self.initial["challenge"] = new_challenge
@@ -33,7 +40,7 @@ class ChallengeMixin(forms.Form):
 class SimpleLoginForm(ChallengeMixin, forms.Form):
     """Basic login form, that can be used as reference for implementation."""
 
-    address = forms.CharField(validators=[validate_monero_address])
+    address = forms.CharField()
     signature = forms.CharField()
 
     error_messages = {
@@ -84,7 +91,7 @@ class SimpleSignUpForm(ChallengeMixin, forms.Form):
     """Basic login form, that can be used as reference for implementation."""
 
     username = forms.CharField()
-    address = forms.CharField(validators=[validate_monero_address])
+    address = forms.CharField()
     signature = forms.CharField()
 
     def __init__(self, request=None, *args, **kwargs):
@@ -93,9 +100,38 @@ class SimpleSignUpForm(ChallengeMixin, forms.Form):
         super().__init__(*args, **kwargs)
         self.request = request
         self.include_challange()
+        self.network = None
 
     def clean_address(self):
+        self.network = None
         value = self.cleaned_data["address"]
+        bitcoin_backend = "django_cryptolock.backends.BitcoinAddressBackend"
+        monero_backend = "django_cryptolock.backends.MoneroAddressBackend"
+
+        if bitcoin_backend in settings.AUTHENTICATION_BACKENDS:
+            try:
+                validate_bitcoin_address(value)
+                self.network = Address.NETWORK_BITCOIN
+            except ValidationError:
+                pass
+
+        if monero_backend in settings.AUTHENTICATION_BACKENDS:
+            try:
+                validate_monero_address(value)
+                self.network = Address.NETWORK_MONERO
+            except ValidationError:
+                pass
+
+        if not self.network:
+            raise forms.ValidationError(_("Invalid address"))
+
         if Address.objects.filter(address=value).exists():
             raise forms.ValidationError(_("This address already exists"))
+
+        return value
+
+    def clean_username(self):
+        value = self.cleaned_data["username"]
+        if get_user_model().objects.filter(username=value).exists():
+            raise forms.ValidationError(_("This username is already taken"))
         return value