+++ django_cryptolock/urls.py
@@ -2,11 +2,11 @@
 from django.conf.urls import url
 from django.views.generic import TemplateView
 
-from .views import MoneroLoginView, MoneroSignUpView
+from .views import CryptoLockLoginView, CryptoLockSignUpView
 
 
 app_name = "django_cryptolock"
 urlpatterns = [
-    url(r"login", MoneroLoginView.as_view(), name="login"),
-    url(r"signup", MoneroSignUpView.as_view(), name="signup"),
+    url(r"login", CryptoLockLoginView.as_view(), name="login"),
+    url(r"signup", CryptoLockSignUpView.as_view(), name="signup"),
 ]