+++ HISTORY.rst
@@ -3,6 +3,12 @@
 History
 -------
 
+(next version)
+++++++++++++++
+
+* Added validation for existing addresses on the signup form
+
+
 0.0.2 (2020-01-08)
 ++++++++++++++++++