+++ HISTORY.rst
@@ -8,6 +8,7 @@ History
 
 * Add validation for existing addresses on the signup form
 * Add rudimentary BitId support
+* Renamed the base auth views to generic names
 
 0.0.2 (2020-01-08)
 ++++++++++++++++++